[FIX] Animals are not colliding with other objects
